+++ ballista/scheduler/src/api/mod.rs
@@ -11,6 +11,8 @@
 // limitations under the License.
 
 #[cfg(feature = "rest-api")]
+mod dto_build;
+#[cfg(feature = "rest-api")]
 mod handlers;
 mod health;
 #[cfg(feature = "rest-api")]
